feat: 커서 기반 페이지네이션 레포지토리 구현

+++ backend/baton/src/main/java/touch/baton/domain/runnerpost/controller/RunnerPostTempController.java
@@ -21,7 +21,7 @@ public class RunnerPostTempController {
     @GetMapping("/temp")
     public ResponseEntity<List<RunnerPostResponse.Simple>> read(
             @RequestParam("cursor") final Long cursor,
-            @RequestParam("limit") final int limit,
+            @RequestParam("limit") final Long limit,
             @RequestParam("reviewStatus") final ReviewStatus reviewStatus
     ) {
         final List<RunnerPost> runnerPosts = runnerPostNativeService.findNative(cursor, limit, reviewStatus);
